reviews_poc/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2
groq==0.4.1
python-dateutil==2.8.2
uvloop==0.19.0
